Update VertexBuffer
Replace textureID variable type from float to int

+++ TSDV-WaveEngine/src/VertexData.h
@@ -25,7 +25,7 @@ WAVEEXPORT struct VertexData
 
 	Vector2 textureCordinates;
 
-	float textureID = 0;
+	int textureID = 0;
 
 	WAVEEXPORT VertexData();
 	WAVEEXPORT VertexData(float v1, float v2, float v3, float r, float g, float b, float a);
